requirements.txt
fastapi
uvicorn
pdfplumber
pytesseract
pillow
python-multipart
langchain
openai
sqlalchemy